Add support for building and location details page
[BPHH-2442]
[BPHH-2265]

- add seeds for jurisdiction HDD and use to provide default value for HDD
  and climate zone on part 3 checklist form
- add UI for building and location details page on part 3 checklist form
  along with necessary API updates

+++ db/migrate/20241030222657_create_part_3_step_code_tables.rb
@@ -2,7 +2,7 @@ class CreatePart3StepCodeTables < ActiveRecord::Migration[7.1]
   def change
     create_table :part_3_step_code_checklists, id: :uuid do |t|
       t.references :step_code, foreign_key: :true, type: :uuid
-      t.decimal :building_height
+      t.integer :building_height
       t.integer :building_code_version
       t.integer :heating_degree_days
       t.integer :climate_zone
